// upvertice-ui/src/app/services/models/page-response-supplier-transaction-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SupplierTransactionResponse } from '../models/supplier-transaction-response';
export interface PageResponseSupplierTransactionResponse {
  content?: Array<SupplierTransactionResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}