// upvertice-ui/src/app/services/models/supplier-transaction-multipart-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Multipart form request for updating supplier transaction with request and images
 */
export interface SupplierTransactionMultipartRequest {

  /**
   * List of images or proofs related to the supplier transaction
   */
  images?: Array<Blob>;

  /**
   * Supplier Transaction update request as JSON string
   */
  request?: string;
}